Cleanup some utils
