+++ core/src/main/java/com/afollestad/materialdialogs/internal/list/MultiChoiceDialogAdapter.kt
@@ -30,9 +30,9 @@ import com.afollestad.materialdialogs.list.MultiChoiceListener
 import com.afollestad.materialdialogs.list.getItemSelector
 import com.afollestad.materialdialogs.utils.appendAll
 import com.afollestad.materialdialogs.utils.inflate
-import com.afollestad.materialdialogs.utils.maybeSetTextColor
 import com.afollestad.materialdialogs.utils.pullIndices
 import com.afollestad.materialdialogs.utils.removeAll
+import com.afollestad.materialdialogs.utils.MDUtil.maybeSetTextColor
 
 /** @author Aidan Follestad (afollestad) */
 internal class MultiChoiceViewHolder(
