+++ core/src/main/java/com/afollestad/materialdialogs/internal/list/PlainListDialogAdapter.kt
@@ -26,7 +26,7 @@ import com.afollestad.materialdialogs.actions.hasActionButtons
 import com.afollestad.materialdialogs.list.ItemListener
 import com.afollestad.materialdialogs.list.getItemSelector
 import com.afollestad.materialdialogs.utils.inflate
-import com.afollestad.materialdialogs.utils.maybeSetTextColor
+import com.afollestad.materialdialogs.utils.MDUtil.maybeSetTextColor
 
 private const val KEY_ACTIVATED_INDEX = "activated_index"
