+++ core/src/main/java/com/afollestad/materialdialogs/utils/ColorExt.kt
@@ -15,8 +15,6 @@
  */
 package com.afollestad.materialdialogs.utils
 
-import android.content.Context
-import android.widget.TextView
 import androidx.annotation.AttrRes
 import androidx.annotation.ColorInt
 import androidx.annotation.ColorRes
@@ -27,14 +25,3 @@ import com.afollestad.materialdialogs.utils.MDUtil.resolveColor
   @ColorRes res: Int? = null,
   @AttrRes attr: Int? = null
 ): Int = resolveColor(windowContext, res, attr)
-
-internal fun TextView?.maybeSetTextColor(
-  context: Context,
-  @AttrRes attrRes: Int?
-) {
-  if (attrRes == null) return
-  val color = resolveColor(context, attr = attrRes)
-  if (color != 0) {
-    this?.setTextColor(color)
-  }
-}
