+++ core/src/main/java/com/afollestad/materialdialogs/utils/DialogExt.kt
@@ -38,6 +38,7 @@ import com.afollestad.materialdialogs.checkbox.getCheckBoxPrompt
 import com.afollestad.materialdialogs.customview.CUSTOM_VIEW_NO_HORIZONTAL_PADDING
 import com.afollestad.materialdialogs.utils.MDUtil.resolveDrawable
 import com.afollestad.materialdialogs.utils.MDUtil.resolveString
+import com.afollestad.materialdialogs.utils.MDUtil.maybeSetTextColor
 
 internal fun MaterialDialog.setWindowConstraints() {
   window?.setSoftInputMode(SOFT_INPUT_ADJUST_RESIZE) ?: return
