+++ files/src/main/java/com/afollestad/materialdialogs/files/DialogFolderChooserExt.kt
@@ -28,7 +28,7 @@ import com.afollestad.materialdialogs.customview.customView
 import com.afollestad.materialdialogs.customview.getCustomView
 import com.afollestad.materialdialogs.files.utilext.hasReadStoragePermission
 import com.afollestad.materialdialogs.files.utilext.hasWriteStoragePermission
-import com.afollestad.materialdialogs.files.utilext.maybeSetTextColor
+import com.afollestad.materialdialogs.utils.MDUtil.maybeSetTextColor
 import com.afollestad.materialdialogs.internal.list.DialogRecyclerView
 import java.io.File
